Make switch to non-root user possible
Commit makes it possible to run as a non-root user which makes sense.
Unfortunately, this means everything will run via % su; and in % bash;.
Given the fact there is at least one re-implementation of % su;, there
might be hidden caveats.
However, I came up only with two solutions. This or allow to specify
command to be executed and write wrapper shipped in Docker Image.